feat: clean code & lint
